Remove unnecessary casts
Some of these date back to pre-C89.
* src/buffer.c (backspace_output):
* src/create.c (to_base256, gid_to_chars, major_to_chars)
(minor_to_chars, off_to_chars, time_to_chars, uid_to_chars):
* src/list.c (from_header, tartime):
* src/map.c (owner_map_translate, group_map_translate):
* src/system.c (sys_truncate):
* src/utf8.c (utf8_init):
* src/xattrs.c (acls_one_line):
* src/xheader.c (xheader_string_end):
Remove casts.
* src/create.c (uintmax_to_chars): Remove.  All uses removed.
(simple_finish_header): Use to_octal.

+++ src/utf8.c
@@ -54,14 +54,14 @@ static iconv_t conv_desc[2] = { (iconv_t) -1, (iconv_t) -1 };
 static iconv_t
 utf8_init (bool to_utf)
 {
-  if (conv_desc[(int) to_utf] == (iconv_t) -1)
+  if (conv_desc[to_utf] == (iconv_t) -1)
     {
       if (to_utf)
-	conv_desc[(int) to_utf] = iconv_open ("UTF-8", locale_charset ());
+	conv_desc[to_utf] = iconv_open ("UTF-8", locale_charset ());
       else
-	conv_desc[(int) to_utf] = iconv_open (locale_charset (), "UTF-8");
+	conv_desc[to_utf] = iconv_open (locale_charset (), "UTF-8");
     }
-  return conv_desc[(int) to_utf];
+  return conv_desc[to_utf];
 }
 
 bool
